services/appointments/requirements.txt
# API and web framework
fastapi
uvicorn
pydantic
pydantic-settings
email-validator

# Database
pymongo

# Authentication and security
python-jose
passlib
python-multipart
bcrypt
httpx

# Messaging
pika
aio_pika

# Caching
redis
aioredis

# Monitoring and instrumentation
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-instrumentation
prometheus-client
python-consul

# Utilities
python-dotenv
python-dateutil
tenacity
structlog
black
isort
pytest
pytest-asyncio